Rename ClassRegistryEntry to RegistryMetadata and delete base.py
Move the shared registry metadata base out of base.py into its own
registry_metadata.py module, renamed from ClassRegistryEntry to the clearer
RegistryMetadata. All six component metadata classes plus scenario/initializer
description helpers now extend/reference it from the new module.

With that moved, base.py held only dead weight: RegistryProtocol had no
production implementers after the legacy-registry deletion, and _matches_filters/
_get_metadata_value were byte-identical duplicates of the copies in registry.py.
instance_registry now uses registry.py's _matches_filters, RegistryProtocol is
dropped from the public API (and the registry doc updated), and base.py is deleted.

Also trim discovery.py: discover_in_directory is still used by the initializer
registry, but discover_in_package and discover_subclasses_in_loaded_modules were
unused exports and are removed.
