+++ pyrit/registry/__init__.py
@@ -3,7 +3,6 @@
 
 """Registry module for PyRIT class and object registries."""
 
-from pyrit.registry.base import RegistryProtocol
 from pyrit.registry.components import (
     AttackTechniqueMetadata,
     AttackTechniqueRegistry,
@@ -18,11 +17,7 @@ from pyrit.registry.components import (
     TargetMetadata,
     TargetRegistry,
 )
-from pyrit.registry.discovery import (
-    discover_in_directory,
-    discover_in_package,
-    discover_subclasses_in_loaded_modules,
-)
+from pyrit.registry.discovery import discover_in_directory
 from pyrit.registry.instance_registry import (
     DefaultInstanceRegistry,
     InstanceRegistry,
@@ -30,6 +25,7 @@ from pyrit.registry.instance_registry import (
     SupportsInstances,
 )
 from pyrit.registry.registry import Registry
+from pyrit.registry.registry_metadata import RegistryMetadata
 from pyrit.registry.tag_query import TagQuery
 
 __all__ = [
@@ -40,14 +36,12 @@ __all__ = [
     "DefaultInstanceRegistry",
     "InstanceRegistry",
     "Registry",
+    "RegistryMetadata",
     "SupportsInstances",
     "discover_in_directory",
-    "discover_in_package",
-    "discover_subclasses_in_loaded_modules",
     "InitializerMetadata",
     "InitializerRegistry",
     "RegistryEntry",
-    "RegistryProtocol",
     "ScenarioMetadata",
     "ScenarioRegistry",
     "ScorerRegistry",
