--- pyrit/registry/base.py
@@ -1,239 +1,0 @@
-# Copyright (c) Microsoft Corporation.
-# Licensed under the MIT license.
-
-"""
-Shared base types for PyRIT registries.
-
-This module contains types shared between class registries (which store type[T])
-and object registries (which store T instances).
-"""
-
-from __future__ import annotations
-
-import inspect
-from dataclasses import dataclass, field
-from typing import TYPE_CHECKING, Any, Protocol, TypeVar, runtime_checkable
-
-if TYPE_CHECKING:
-    from collections.abc import Iterator, Mapping
-
-    from typing_extensions import Self
-
-    from pyrit.models.parameter import Parameter
-
-# Type variable for metadata (invariant for Protocol compatibility)
-MetadataT = TypeVar("MetadataT")
-
-
-@dataclass(frozen=True)
-class ClassRegistryEntry:
-    """
-    Minimal base for class-level registry metadata.
-
-    Provides the common fields every registry metadata type needs for display,
-    lookup, and filtering in class registries.
-
-    Attributes:
-        class_name (str): Python class name (e.g., "ContentHarmsScenario").
-        class_module (str): Full module path (e.g., "pyrit.scenario.scenarios.content_harms").
-        class_description (str): Human-readable description, typically from the class docstring.
-        registry_name (str): The suffix-stripped snake_case key used in the registry
-            (e.g., "content_harms" for ContentHarmsScenario).
-        parameters (tuple[Parameter, ...]): The derived build contract for the class.
-            Buildable registries (e.g. converters) populate this from the constructor
-            signature; scenarios/initializers use their own ``supported_parameters``
-            today and will migrate to this unified shape.
-        class_attributes (Mapping[str, Any]): Values sourced from class attributes
-            (declared on the identifier via ``Param.ClassAttr``), letting the entry
-            describe class-level facts — e.g. a converter's supported input/output
-            types — without constructing an instance. Empty for entries with none.
-    """
-
-    class_name: str
-    class_module: str
-    class_description: str = ""
-    registry_name: str = ""
-    parameters: tuple[Parameter, ...] = field(kw_only=True, default=())
-    class_attributes: Mapping[str, Any] = field(kw_only=True, default_factory=dict)
-
-    @staticmethod
-    def description_from_docstring(cls: type, *, fallback: str = "") -> str:
-        """
-        Extract a normalized description from a class docstring.
-
-        Collapses all whitespace into single spaces. Returns fallback if
-        no docstring is present or the docstring is empty after cleaning.
-
-        Returns:
-            str: The cleaned docstring or the fallback value.
-        """
-        doc = cls.__doc__ or ""
-        cleaned = " ".join(doc.split())
-        return cleaned or fallback
-
-    @staticmethod
-    def summary_from_docstring(cls: type) -> str:
-        """
-        Extract a short summary from the first paragraph of a class docstring.
-
-        Uses the class's own docstring only (never an inherited one), normalizes
-        indentation, and collapses the first paragraph's whitespace onto one line.
-        Empty when the class has no docstring. This is the catalog-display
-        counterpart to ``description_from_docstring`` (which collapses the whole
-        docstring); buildable registries populate ``class_description`` from this
-        first-paragraph form.
-
-        Returns:
-            str: The first-paragraph summary, or "" when there is no docstring.
-        """
-        raw = cls.__doc__
-        if not raw:
-            return ""
-        first_paragraph = inspect.cleandoc(raw).split("\n\n", 1)[0]
-        return " ".join(first_paragraph.split())
-
-
-@runtime_checkable
-class RegistryProtocol(Protocol[MetadataT]):
-    """
-    Protocol defining the common interface for all registries.
-
-    Registries implement this interface, enabling code that works with any
-    registry type.
-
-    Type Parameters:
-        MetadataT: The metadata dataclass type (e.g., ScenarioMetadata).
-    """
-
-    @classmethod
-    def get_registry_singleton(cls) -> Self:
-        """Get the singleton instance of this registry."""
-        ...
-
-    @classmethod
-    def reset_instance(cls) -> None:
-        """Reset the singleton instance."""
-        ...
-
-    def get_names(self) -> list[str]:
-        """Get a sorted list of all registered names."""
-        ...
-
-    def list_metadata(
-        self,
-        *,
-        include_filters: dict[str, Any] | None = None,
-        exclude_filters: dict[str, Any] | None = None,
-    ) -> list[MetadataT]:
-        """
-        List metadata for all registered items, optionally filtered.
-
-        Args:
-            include_filters: Optional dict of filters that items must match.
-                Keys are metadata property names, values are the filter criteria.
-                All filters must match (AND logic).
-            exclude_filters: Optional dict of filters that items must NOT match.
-                Keys are metadata property names, values are the filter criteria.
-                Any matching filter excludes the item.
-
-        Returns:
-            List of metadata describing each registered item.
-        """
-        ...
-
-    def __contains__(self, name: str) -> bool:
-        """Check if a name is registered."""
-        ...
-
-    def __len__(self) -> int:
-        """Get the count of registered items."""
-        ...
-
-    def __iter__(self) -> Iterator[str]:
-        """Iterate over registered names."""
-        ...
-
-
-def _get_metadata_value(metadata: Any, key: str) -> tuple[bool, Any]:
-    """
-    Get a value from a metadata object by key.
-
-    Checks direct attributes first, then falls back to the ``params`` dict
-    (used by ComponentIdentifier). Returns a (found, value) tuple.
-
-    Args:
-        metadata: The metadata object to look up.
-        key (str): The attribute or params key to find.
-
-    Returns:
-        tuple: (True, value) if found, (False, None) otherwise.
-    """
-    if hasattr(metadata, key):
-        return True, getattr(metadata, key)
-
-    # Fall back to params dict (for ComponentIdentifier)
-    params = getattr(metadata, "params", None)
-    if isinstance(params, dict) and key in params:
-        return True, params[key]
-
-    return False, None
-
-
-def _matches_filters(
-    metadata: Any,
-    *,
-    include_filters: dict[str, Any] | None = None,
-    exclude_filters: dict[str, Any] | None = None,
-) -> bool:
-    """
-    Check if a metadata object matches all provided filters.
-
-    Supports filtering on any property of the metadata dataclass or on keys
-    inside the ``params`` dict (for ComponentIdentifier metadata):
-    - For simple types (str, int, bool): exact match comparison
-    - For sequence types (list, tuple): checks if filter value is contained in the sequence
-
-    Items must match ALL include_filters (AND logic) and must NOT match ANY exclude_filters.
-
-    Args:
-        metadata: The metadata dataclass instance to check.
-        include_filters: Optional dict of filters that must ALL match.
-            Keys are metadata property names or params keys, values are the filter criteria.
-        exclude_filters: Optional dict of filters that must ALL NOT match.
-            Keys are metadata property names or params keys, values are the filter criteria.
-
-    Returns:
-        True if all include_filters match and no exclude_filters match, False otherwise.
-    """
-    # Check include filters - all must match
-    if include_filters:
-        for key, filter_value in include_filters.items():
-            found, actual_value = _get_metadata_value(metadata, key)
-            if not found:
-                return False
-
-            # Handle sequence types - check if filter value is in the sequence
-            if isinstance(actual_value, (list, tuple)):
-                if filter_value not in actual_value:
-                    return False
-            # Simple exact match for other types
-            elif actual_value != filter_value:
-                return False
-
-    # Check exclude filters - none must match
-    if exclude_filters:
-        for key, filter_value in exclude_filters.items():
-            found, actual_value = _get_metadata_value(metadata, key)
-            if not found:
-                # If the key doesn't exist, it can't match the exclude filter
-                continue
-
-            # Handle sequence types - exclude if filter value is in the sequence
-            if isinstance(actual_value, (list, tuple)):
-                if filter_value in actual_value:
-                    return False
-            # Simple exact match for other types - exclude if it matches
-            elif actual_value == filter_value:
-                return False
-
-    return True
