+++ pyrit/registry/components/scorer_registry.py
@@ -29,9 +29,9 @@ from typing import TYPE_CHECKING
 
 from pyrit.models.identifiers import ScorerIdentifier
 from pyrit.models.parameter import ComponentType
-from pyrit.registry.base import ClassRegistryEntry
 from pyrit.registry.instance_registry import DefaultInstanceRegistry, InstanceRegistry
 from pyrit.registry.registry import Registry
+from pyrit.registry.registry_metadata import RegistryMetadata
 
 if TYPE_CHECKING:
     from types import ModuleType
@@ -40,7 +40,7 @@ if TYPE_CHECKING:
 
 
 @dataclass(frozen=True)
-class ScorerMetadata(ClassRegistryEntry):
+class ScorerMetadata(RegistryMetadata):
     """
     Metadata describing a registered ``Scorer`` class.
