+++ pyrit/setup/initializers/pyrit_initializer.py
@@ -87,9 +87,9 @@ class PyRITInitializer(ABC):
         Returns:
             str: A description of the configuration changes this initializer makes.
         """
-        from pyrit.registry.base import ClassRegistryEntry
+        from pyrit.registry.registry_metadata import RegistryMetadata
 
-        return ClassRegistryEntry.description_from_docstring(self.__class__, fallback=type(self).__name__)
+        return RegistryMetadata.description_from_docstring(self.__class__, fallback=type(self).__name__)
 
     @property
     def required_env_vars(self) -> list[str]:
